Minor changes to DTO properties

+++ models/CheckoutAndLinksDTOs/RequestContent/PaymentLinkRequestContent/TransactionAmount.php
@@ -3,6 +3,6 @@ use Fiserv\models\FiservObject;
 
 class transactionAmount extends FiservObject
 {
-    public int $total;
+    public float $total;
     public string $currency;
 }
